ec/backend/requirements.txt
wheel
dj-database-url
dj-static
Django
python-decouple
gunicorn
psycopg2-binary
mercadopago
djangorestframework
django-cors-headers
